add entity: delivery: DeliveryAssignment; payment: Payment, Refund; review: Review, ReviewResponse, ReviewVote, search: ProductView, SearcHistory; transaction: Statistic, TransactionHistory; voucher: FlashSale, Voucher, VoucherUsage; wishlist: Wishlist, WishlistItem; Notification | refactorentity user/User long -> int

+++ src/main/java/viettech/entity/user/User.java
@@ -137,7 +137,7 @@ public abstract class User {
        GETTERS & SETTERS
        ========================= */
 
-    public Long getUserId() {
+    public int getUserId() {
         return userId;
     }
